+++ angular/src/app/components/create-event/create-event.component.html
@@ -1,6 +1,6 @@
 <div class="govuk-grid-row">
   <div class="govuk-grid-column-two-thirds">
-    <app-form-stepper [caseId]="caseId" [files]="files" (submitEvent)="onSubmit($event)" [pages]="pages"></app-form-stepper>
+    <app-linear-stepper [caseId]="caseId" [files]="files" (submitEvent)="onSubmit($event)" [pages]="pages"></app-linear-stepper>
   </div>
 </div>
